+++ WHATSNEW.md
@@ -8,9 +8,13 @@
 - CLI command "ems get workgroups" now displays group id in table
 - Tool "ems get workgroups" has changed its return format
 - Massive optimizations in cli help tests
+- Changed typing for list, dict, set and tuple
 
 ### Fixed
 
 - Fixed some typos
 
 ### Removed
+
+- Remove support for Python3.8
+- Remove upper version bounds for external dependencies